+++ src/dlab/definition.js
@@ -2,13 +2,13 @@ import convertDlabToLab from './convertDlabToLab';
 import convertDlabToRgb from './convertDlabToRgb';
 import convertLabToDlab from './convertLabToDlab';
 import convertRgbToDlab from './convertRgbToDlab';
-import interpolateNumber from '../interpolate/interpolateNumber';
-import interpolateAlpha from '../interpolate/interpolateAlpha';
-import interpolateFunctionLinear from '../interpolate/interpolateFunctionLinear';
+import interpolateNumber from '../interpolate/number';
+import interpolateAlpha from '../interpolate/alpha';
+import interpolateLinear from '../interpolate/linear';
 
 export default {
 	mode: 'dlab',
-	output: { 
+	output: {
 		lab: convertDlabToLab,
 		rgb: convertDlabToRgb
 	},
@@ -18,9 +18,9 @@ export default {
 	},
 	channels: ['l', 'a', 'b', 'alpha'],
 	interpolate: {
-		l: interpolateFunctionLinear(interpolateNumber()),
-		a: interpolateFunctionLinear(interpolateNumber()),
-		b: interpolateFunctionLinear(interpolateNumber()),
-		alpha: interpolateFunctionLinear(interpolateAlpha())
+		l: interpolateLinear(interpolateNumber()),
+		a: interpolateLinear(interpolateNumber()),
+		b: interpolateLinear(interpolateNumber()),
+		alpha: interpolateLinear(interpolateAlpha())
 	}
 };
