+++ src/dlch/definition.js
@@ -6,14 +6,14 @@ import convertLabToDlch from './convertLabToDlch';
 import convertDlabToDlch from './convertDlabToDlch';
 import convertRgbToDlch from './convertRgbToDlch';
 
-import interpolateNumber from '../interpolate/interpolateNumber';
-import interpolateAlpha from '../interpolate/interpolateAlpha';
-import interpolateHue from '../interpolate/interpolateHue';
-import interpolateFunctionLinear from '../interpolate/interpolateFunctionLinear';
+import interpolateNumber from '../interpolate/number';
+import interpolateAlpha from '../interpolate/alpha';
+import interpolateHue from '../interpolate/hue';
+import interpolateLinear from '../interpolate/linear';
 
 export default {
 	mode: 'dlch',
-	output: { 
+	output: {
 		lab: convertDlchToLab,
 		dlab: convertDlchToDlab,
 		rgb: convertDlchToRgb
@@ -25,9 +25,9 @@ export default {
 	},
 	channels: ['l', 'c', 'h', 'alpha'],
 	interpolate: {
-		l: interpolateFunctionLinear(interpolateNumber()),
-		c: interpolateFunctionLinear(interpolateNumber()),
-		h: interpolateFunctionLinear(interpolateHue()),
-		alpha: interpolateFunctionLinear(interpolateAlpha())
+		l: interpolateLinear(interpolateNumber()),
+		c: interpolateLinear(interpolateNumber()),
+		h: interpolateLinear(interpolateHue()),
+		alpha: interpolateLinear(interpolateAlpha())
 	}
 };
