+++ src/hsi/definition.js
@@ -1,23 +1,23 @@
 import convertHsiToRgb from './convertHsiToRgb';
 import convertRgbToHsi from './convertRgbToHsi';
-import interpolateNumber from '../interpolate/interpolateNumber';
-import interpolateAlpha from '../interpolate/interpolateAlpha';
-import interpolateHue from '../interpolate/interpolateHue';
-import interpolateFunctionLinear from '../interpolate/interpolateFunctionLinear';
+import interpolateNumber from '../interpolate/number';
+import interpolateAlpha from '../interpolate/alpha';
+import interpolateHue from '../interpolate/hue';
+import interpolateLinear from '../interpolate/linear';
 
 export default {
 	mode: 'hsi',
-	output: { 
-		rgb: convertHsiToRgb 
+	output: {
+		rgb: convertHsiToRgb
 	},
 	input: {
 		rgb: convertRgbToHsi
 	},
 	channels: ['h', 's', 'i', 'alpha'],
 	interpolate: {
-		h: interpolateFunctionLinear(interpolateHue()),
-		s: interpolateFunctionLinear(interpolateNumber()),
-		i: interpolateFunctionLinear(interpolateNumber()),
-		alpha: interpolateFunctionLinear(interpolateAlpha())
+		h: interpolateLinear(interpolateHue()),
+		s: interpolateLinear(interpolateNumber()),
+		i: interpolateLinear(interpolateNumber()),
+		alpha: interpolateLinear(interpolateAlpha())
 	}
-}
+};
