+++ src/hsl/definition.js
@@ -1,25 +1,25 @@
 import convertHslToRgb from './convertHslToRgb';
 import convertRgbToHsl from './convertRgbToHsl';
 import parseHsl from './parseHsl';
-import interpolateNumber from '../interpolate/interpolateNumber';
-import interpolateAlpha from '../interpolate/interpolateAlpha';
-import interpolateHue from '../interpolate/interpolateHue';
-import interpolateFunctionLinear from '../interpolate/interpolateFunctionLinear';
+import interpolateNumber from '../interpolate/number';
+import interpolateAlpha from '../interpolate/alpha';
+import interpolateHue from '../interpolate/hue';
+import interpolateLinear from '../interpolate/linear';
 
 export default {
 	mode: 'hsl',
-	output: { 
-		rgb: convertHslToRgb 
+	output: {
+		rgb: convertHslToRgb
 	},
 	input: {
 		rgb: convertRgbToHsl
 	},
 	channels: ['h', 's', 'l', 'alpha'],
-	parsers: [ parseHsl ],
+	parsers: [parseHsl],
 	interpolate: {
-		h: interpolateFunctionLinear(interpolateHue()),
-		s: interpolateFunctionLinear(interpolateNumber()),
-		l: interpolateFunctionLinear(interpolateNumber()),
-		alpha: interpolateFunctionLinear(interpolateAlpha())
+		h: interpolateLinear(interpolateHue()),
+		s: interpolateLinear(interpolateNumber()),
+		l: interpolateLinear(interpolateNumber()),
+		alpha: interpolateLinear(interpolateAlpha())
 	}
 };
