+++ src/hwb/definition.js
@@ -1,25 +1,25 @@
 import convertHwbToRgb from './convertHwbToRgb';
 import convertRgbToHwb from './convertRgbToHwb';
 import parseHwb from './parseHwb';
-import interpolateNumber from '../interpolate/interpolateNumber';
-import interpolateAlpha from '../interpolate/interpolateAlpha';
-import interpolateHue from '../interpolate/interpolateHue';
-import interpolateFunctionLinear from '../interpolate/interpolateFunctionLinear';
+import interpolateNumber from '../interpolate/number';
+import interpolateAlpha from '../interpolate/alpha';
+import interpolateHue from '../interpolate/hue';
+import interpolateLinear from '../interpolate/linear';
 
 export default {
 	mode: 'hwb',
-	output: { 
-		rgb: convertHwbToRgb 
+	output: {
+		rgb: convertHwbToRgb
 	},
 	input: {
 		rgb: convertRgbToHwb
 	},
 	channels: ['h', 'w', 'b', 'alpha'],
-	parsers: [ parseHwb ],
+	parsers: [parseHwb],
 	interpolate: {
-		h: interpolateFunctionLinear(interpolateHue()),
-		w: interpolateFunctionLinear(interpolateNumber()),
-		b: interpolateFunctionLinear(interpolateNumber()),
-		alpha: interpolateFunctionLinear(interpolateAlpha())
+		h: interpolateLinear(interpolateHue()),
+		w: interpolateLinear(interpolateNumber()),
+		b: interpolateLinear(interpolateNumber()),
+		alpha: interpolateLinear(interpolateAlpha())
 	}
 };
