+++ src/interpolate/cosine.js
@@ -5,15 +5,14 @@
 
 import identity from '../util/identity';
 
-export default (normalize = identity, γ = 1) => 
-	(arr, t) => {
-
+export default (γ = 1) => (arr, normalize) => {
+	return t => {
 		t = Math.pow(t, γ);
 
-		let cls = t * (arr.length - 1), 
+		let cls = t * (arr.length - 1),
 			idx = Math.floor(cls),
-			a = arr[idx], 
-			b = arr[idx + 1], 
+			a = arr[idx],
+			b = arr[idx + 1],
 			t0 = cls - idx;
 
 		let values = normalize([a, b], t0);
@@ -21,8 +20,9 @@ export default (normalize = identity, γ = 1) =>
 			a = values[0];
 			b = values[1];
 			let c = (1 - Math.cos(t0 * Math.PI)) / 2;
-			return (a * (1 - c) + b * c);
+			return a * (1 - c) + b * c;
 		} else {
 			return values;
-		} 
+		}
 	};
+};
