+++ src/lch/definition.js
@@ -3,10 +3,10 @@ import convertLchToLab from './convertLchToLab';
 import convertLchToRgb from './convertLchToRgb';
 import convertRgbToLch from './convertRgbToLch';
 import parseLch from './parseLch';
-import interpolateNumber from '../interpolate/interpolateNumber';
-import interpolateAlpha from '../interpolate/interpolateAlpha';
-import interpolateHue from '../interpolate/interpolateHue';
-import interpolateFunctionLinear from '../interpolate/interpolateFunctionLinear';
+import interpolateNumber from '../interpolate/number';
+import interpolateAlpha from '../interpolate/alpha';
+import interpolateHue from '../interpolate/hue';
+import interpolateLinear from '../interpolate/linear';
 
 export default {
 	mode: 'lch',
@@ -14,16 +14,16 @@ export default {
 		lab: convertLchToLab,
 		rgb: convertLchToRgb
 	},
-	input: { 
+	input: {
 		rgb: convertRgbToLch,
 		lch: convertLabToLch
 	},
 	channels: ['l', 'c', 'h', 'alpha'],
-	parsers: [ parseLch ],
+	parsers: [parseLch],
 	interpolate: {
-		h: interpolateFunctionLinear(interpolateHue()),
-		c: interpolateFunctionLinear(interpolateNumber()),
-		l: interpolateFunctionLinear(interpolateNumber()),
-		alpha: interpolateFunctionLinear(interpolateAlpha())
+		h: interpolateLinear(interpolateHue()),
+		c: interpolateLinear(interpolateNumber()),
+		l: interpolateLinear(interpolateNumber()),
+		alpha: interpolateLinear(interpolateAlpha())
 	}
 };
